+++ app/src/main/java/devchallenge/android/radiotplayer/repository/modelspec/PodcastInfoSpec.java
@@ -28,6 +28,6 @@ class PodcastInfoSpec {
     int length;
 
     @ColumnSpec(defaultValue="0")
-    long fileSize;
+    int fileSize;
 
 }
